[ADD] now containers can have a requirements-firsttime.txt to control packages to install only firsttime container starts and requirements-app.txt for packages to install every time container ups, usefull for develepment using packages under development (not stable) [UPD] rename image from ege_django_base to ege_base

+++ startup.sh
@@ -1,7 +1,20 @@
 #!/bin/sh
-pip install -r requirements-app.txt
 
-if [ "True" = "$DJANGO_DEBUG" ]; then
+
+if [[ ! -s "/firsttime.lock" ]]; then
+    if [[ -s "requirements-firsttime.txt" ]]; then
+        pip install -r requirements-firsttime.txt
+        date > /firsttime.lock
+    fi
+fi
+
+
+if [[ -s "requirements-app.txt" ]]; then
+    pip install -r requirements-app.txt
+fi
+
+
+if [[ "True" = "$DJANGO_DEBUG" ]]; then
     python3 manage.py runserver 0.0.0.0:8000
 else
     gunicorn \
